{"schema":"decodex/commit/1","summary":"Modularize frozen edit history runtime","authority":"manual"}
